Create Colors instances
With adding comparisons between cold and warm colors, it made sense to
have that logic live in just one place - the Colors class.  This way we
can have that class handle comparisons between different colors.

Note that this won't work right now, because nothing is instantiating
Colors.  That is the next improvement to be made.