Add fields to Scenario model
This adds four fields to the Scenario model, only one of which is used
at this time: the requirement count.  That is now included so we can
show it in the index and people can get an idea of how hard certain
games are.  The other fields are in preparation for future changes, when
we include the ability to rate scenarios.