Fix random color append
The previous code appended the "warm" and "cool" colors to the total
list of colors - which meant that both A) those were now available as
random colors when selecting colors for rooms or furnishings, and B)
over time, as the method was called, those colors would start to be over
represented in the list of colors, so everything would become "warm" or
"cool".

Instead, just make a new array that includes both the original colors
and "warm" and "cool", and select from that.