Add rule for an exact single furnishing
This required including a way to get a random furnishing type (ie:
`Lamp`, `Curio`, or `WallHanging`), without getting an
`EmptyFurnishing`.  It also required having a way to look up a specific
furnishing type for a room.

I think there might be some refactoring that needs to be done here.
Needing to get specifically the Lamp for a room is going to be more
common with some of the later rules, and possibly the Foundational Rules
I'm thinking of implementing.  It may be that this will get refactored
as I make the `tokens` of `Room` into three separate database
properties which point to the individual features of the room.