Rework scenario list for mobile responsiveness
The big change here is to make the scenario list respond to a small
screen by becoming a list instead of a table.

To make this work, I also needed to adjust some of the other settings.
One great improvement was removing the specific px setting for the table
row height - that was problematic for different display resolutions.