Use `singular_rule` rather than `nil`
When creating a computed rule for which there can only be one in the
house, we need `random_feature` to somehow indicate that the rule has
been taken, and any other attempt to use this rule will be a duplicate
of the others.  At first I used `nil` to indicate that there was no
random feature for the rule.  However, that isn't very descriptive and
in other parts of the code it would not be clear why this feature was
returning `nil`.  By making the name a consistent string it is more
clear that this is an intended value.