Aspect must be most frequent rule
Add a new rule, which looks to see what aspect (color, style, or
furnishing) is most frequent in the house, and then requires that no
other aspect appears more frequently than that aspect.

To support this, new methods were added to Room and House, and
Furnishing was chosen to be where the random selection of an aspect
type.

While working on this, it was taking over a minute to run any tests, and
most of that time was spend on loading the files.  This was getting to
be really annoying.  I enabled `spring` for rspec, so that I could just
use that for testing - and it helps a lot.  Tests are down to about 5
seconds total.