Fix AspectMustBeMostFrequent pluralization
I noticed that the pluralization was off on this rule.  The stub method
was returning "lamps", but the real method would return "lamp".  That
means that the pluralization of the output was off, once fixed, so we
needed to add a check for pluralization to the rule.

Note that we only need to pluralize the furnishings.  Styles (modern,
etc) and colors (blue, etc) are not pluralizeds.