Add color spans to the color text
Whenever we are showing the text for a color, we can now include a
`span` element with the class indicating which color to display.

A note on the colors: Most I just used the default, but the default
yellow is really hard to read.  Instead I used Mikado Yellow.

Another note: given how the application stores rules, older scenarios
that were created before this change will not have rules with the
updated color text.